requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
amqp==2.1.4               # via kombu
anymarkup-core==0.7.0     # via anymarkup
anymarkup==0.7.0
beautifulsoup4==4.5.3
BigQuery-Python==1.10.0
billiard==3.5.0.2         # via celery
boto3==1.4.4
boto==2.45.0
botocore==1.5.0           # via boto3, s3transfer
celery==4.0.2             # via selinon
codegen==1.0              # via selinonlib
configobj==5.0.6          # via anymarkup
dill==0.2.5               # via selinonlib
docutils==0.13.1          # via botocore
google-api-python-client==1.5.5  # via bigquery-python
graphviz==0.5.1           # via selinonlib
httplib2==0.9.2           # via bigquery-python, google-api-python-client, oauth2client
jmespath==0.9.0           # via boto3, botocore
jsl==0.2.4
json5==0.2.4              # via anymarkup
jsonschema==2.5.1
kombu==4.0.2              # via celery
lxml==3.6.4
oauth2client==4.0.0       # via google-api-python-client
psycopg2==2.6.2
pyasn1-modules==0.0.8     # via oauth2client
pyasn1==0.1.9             # via oauth2client, pyasn1-modules, rsa
PyGithub==1.29
pymongo==3.3.1
python-dateutil==2.5.3    # via bigquery-python, botocore
pytz==2016.7              # via celery
pyyaml==3.12              # via anymarkup, selinonlib
redis==2.10.5             # via selinon
requests==2.12.4
rsa==3.4.2                # via oauth2client
s3transfer==0.1.10        # via boto3
selinon==0.1.0rc3
selinonlib==0.1.0rc3
six==1.10.0               # via anymarkup-core, configobj, google-api-python-client, oauth2client, python-dateutil
sqlalchemy==1.1.3
toml==0.9.2               # via anymarkup
unidiff==0.5.2
uritemplate==3.0.0        # via google-api-python-client
vine==1.1.3               # via amqp
xmltodict==0.10.2         # via anymarkup
